Split propagation.HTTPFormat implementations
Fixes #358.
